refactor(cloud-agent): keep billing errors local
