+++ packages/container-usage/src/contracts.ts
@@ -133,17 +133,6 @@ export const recordStartFailureCodeSchema = z.enum([
 export type RecordStartFailureCode = z.infer<typeof recordStartFailureCodeSchema>;
 
 /** Customer-safe billing admission failure carried across Worker and client boundaries. */
-export const customerBillingFailureSchema = z
-  .object({
-    code: z.enum(['INSUFFICIENT_CREDITS', 'COMPUTE_STOPPING', 'BILLING_UNAVAILABLE']),
-    payer: z.object({ type: z.enum(['user', 'org']), id: z.string().min(1) }).strict(),
-    retryable: z.boolean(),
-    remainingMicrodollars: z.number().int().nonnegative().optional(),
-    minimumRequiredMicrodollars: z.number().int().nonnegative().optional(),
-  })
-  .strict();
-export type CustomerBillingFailure = z.infer<typeof customerBillingFailureSchema>;
-
 const recordStartFailureSchema = z.discriminatedUnion('code', [
   z
     .object({
