+++ services/cloud-agent-next/src/execution/types.ts
@@ -11,7 +11,6 @@ import type { AgentMode } from '../schema.js';
 import type { Attachments } from '../router/schemas.js';
 import type { SessionMetadata } from '../persistence/session-metadata.js';
 import type { CloudAgentSessionState } from '../persistence/types.js';
-import type { CustomerBillingFailure } from '@kilocode/container-usage/contracts';
 
 // ---------------------------------------------------------------------------
 // Execution Modes
@@ -231,7 +230,14 @@ export type RetryableResultCode =
 
 export type PermanentDeliveryResultCode = 'SANDBOX_CAPABILITY_UNAVAILABLE';
 
-export type { CustomerBillingFailure } from '@kilocode/container-usage/contracts';
+// Keep aligned with the customer-safe schemas in trpc-error.ts and cloud-agent-sdk.
+export type CustomerBillingFailure = {
+  code: 'INSUFFICIENT_CREDITS' | 'COMPUTE_STOPPING' | 'BILLING_UNAVAILABLE';
+  payer: { type: 'user' | 'org'; id: string };
+  retryable: boolean;
+  remainingMicrodollars?: number;
+  minimumRequiredMicrodollars?: number;
+};
 
 export type AdmissionFailure = {
   success: false;
